Rename axolgo-aws to axolgo-cloud
