+++ README.md
@@ -1,20 +1,20 @@
-# axolgo-aws, the Axolotl AWS Library in Golang
-This is the AWS library of the Axolotl series in Golang. The basic calls like loading AWS configuration is handled for you. The AWS calls are also wrapped so you can focus on building your business logic. The interfaces and functions are designed to be more friendly to command line use cases.
+# axolgo-cloud, the Axolotl Cloud Library in Golang
+This is the cloud library of the Axolotl series in Golang. It is a middle layer for the application or CLI to use AWS and GCP. The basic calls like loading AWS configuration is handled for you. The difficult part of using Google API in Golang is handled for you. The client calls are also wrapped so you can focus on building your business logic. The interfaces and functions are designed to be more friendly to command line use cases.
 
 ## Use it with your Go module
 To add as dependency for your package or upgrade to the latest version:
 ```
-go get github.com/tchiunam/axolgo-aws
+go get github.com/tchiunam/axolgo-cloud
 ```
 
 To upgrade or downgrade to a specific version:
 ```
-go get github.com/tchiunam/axolgo-aws@v1.2.3
+go get github.com/tchiunam/axolgo-cloud@v1.2.3
 ```
 
 To remove dependency on your module and downgrade modules:
 ```
-go get github.com/tchiunam/axolgo-aws@none
+go get github.com/tchiunam/axolgo-cloud@none
 ```
 
 See 'go help get' or https://golang.org/ref/mod#go-get for details.
