+++ aws/ec2/describe_instances.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
-	"github.com/tchiunam/axolgo-aws/util"
+	"github.com/tchiunam/axolgo-cloud/aws/util"
 )
 
 // EC2DescribeInstancesAPI defines the interface for the DescribeInstances function.
