+++ aws/util/util.go
